requirements.txt
# PyTorch and ML dependencies
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Tokenizer
sentencepiece>=0.1.99

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Model optimization
onnx>=1.15.0
onnxruntime>=1.16.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
jsonschema>=4.20.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Mobile deployment
coremltools>=7.0  # For iOS Core ML export
onnx-tf>=1.10.0   # For Android TFLite export
tensorflow>=2.14.0  # For TFLite conversion
